give each a transceiver a copy of each packet
we need to do this (for now) since, although each transceiver will
filter out rtp packets for the media type they don't want, they'll both
handle rtcp.  if they share a buffer the first transceiver to decrypt it
will modify it and then the other one will try and decrypt the
already-decrypted packet.  once we have a single transceiver we can do
all rtcp processing in a single leg of the pipeline