add an encodings manager to act as a central place for publish all source encoding updates
currently it's just used for ssrc associations, as that's the first
thing we care about getting out to everyone (the send pipeline's
retransmission senders need to know which ssrc to use when sending an
RTX packet, and this comes from the original sender's encoding
configuration).

in the future i imagine some (sub)set of payload type/format encodings, receive ssrcs and
header extensions may flow through this as well...but if not then we can
either keep this as is or find something a bit less substantial to
achieve the same end for ssrc associations.