fix audio being broken
we cleared/updates payload types/extensions after processing the
channels in each content, rather than after all contents.  this meant
that whatever PTs/extensions the first channel had would get cleared
when the second one was processed.

even this solution is not bug free (it doesn't handle an IQ which
updates only one of the channels and leaves the other alone) so there's
still more work that needs to be done here.