wire up a basic health check which creates a conference and 2 endpoints.
the ice connection (which appears to have been disabled in the old code
anyway) isn't working because we only ever use single port mode (since
we don't support rtcpmux and a singleport harvester is enabled).  i
think this causes the ice agent to fail because both their origin
address/ports are the same.  forcing non rtcpmux is enough to make
things connect, but it breaks other things since we don't support
non rtcpmux right now.  maybe there's a good way to force certain
transport managers to not use the single port harvesters and we can do that
for health check (or just leave it like this, since the old code
apparently wasn't having them connect anyways)